separate typedefs to 	ypedef.jl
